refactor: extract `getDB` util
