+++ apps/extension/src/commons/utils.ts
@@ -1,8 +1,11 @@
+import defaultKanjiFilterRules from "@/assets/rules/filter.json";
 import defaultSelectorRules from "@/assets/rules/selector.json";
+import { type DBSchema, openDB } from "idb";
 import {
   DisplayMode,
   ExtEvent,
   ExtStorage,
+  type FilterRule,
   FuriganaType,
   type GeneralSettings,
   type MoreSettings,
@@ -97,3 +100,33 @@ export const customRules = storage.defineItem<SelectorRule[]>("local:customRules
   version: 1,
   fallback: defaultSelectorRules,
 });
+
+export const DB = {
+  name: "kanjiFilterDB",
+  version: 1,
+  onlyTable: "kanjiFilterTable",
+} as const;
+interface KanjiFilterDB extends DBSchema {
+  [DB.onlyTable]: {
+    key: string;
+    value: FilterRule;
+  };
+}
+
+export const getKanjiFilterDB = async () => {
+  const db = await openDB<KanjiFilterDB>(DB.name, DB.version, {
+    /**
+     * @param transaction
+     * Don't use `db.transaction(...)`, the upgrade callback will run a version change transaction,
+     * and new transactions can't be created until this transaction ends.
+     */
+    upgrade(db, _, __, transaction) {
+      db.createObjectStore(DB.onlyTable, { keyPath: "kanji" });
+      const store = transaction.objectStore(DB.onlyTable);
+      for (const rule of defaultKanjiFilterRules) {
+        store.put(rule);
+      }
+    },
+  });
+  return db;
+};
