+++ apps/extension/src/entrypoints/background/listeners/onGetKanjiMarksMessage.ts
@@ -1,11 +1,7 @@
-import kuromoji from "@sglkc/kuromoji";
-import { type DBSchema, openDB } from "idb";
-
-import type { FilterRule } from "@/commons/constants";
 import { onMessage } from "@/commons/message";
 import { type KanjiToken, type MojiToken, toKanjiToken } from "@/commons/toKanjiToken";
-
-import defaultKanjiFilterRules from "@/assets/rules/filter.json";
+import { DB, getKanjiFilterDB } from "@/commons/utils";
+import kuromoji from "@sglkc/kuromoji";
 
 interface Tokenizer {
   tokenize: (text: string) => MojiToken[];
@@ -52,38 +48,14 @@ const getTokenizer = async () => {
 export interface KanjiMark extends KanjiToken {
   isFiltered: boolean;
 }
-const DATABASE = {
-  name: "kanjiFilterDB",
-  version: 1,
-  onlyTable: "kanjiFilterTable",
-} as const;
-interface KanjiFilterDB extends DBSchema {
-  [DATABASE.onlyTable]: {
-    key: string;
-    value: FilterRule;
-  };
-}
 
 const kanjiFilterMap: Promise<Map<string, string[]>> | null = null;
 const getKanjiFilterMap = async () => {
   if (kanjiFilterMap) {
     return kanjiFilterMap;
   }
-  const db = await openDB<KanjiFilterDB>(DATABASE.name, DATABASE.version, {
-    /**
-     * @param transaction
-     * Don't use `db.transaction(...)`, the upgrade callback will run a version change transaction,
-     * and new transactions can't be created until this transaction ends.
-     */
-    upgrade(db, _, __, transaction) {
-      db.createObjectStore(DATABASE.onlyTable, { keyPath: "kanji" });
-      const store = transaction.objectStore(DATABASE.onlyTable);
-      for (const rule of defaultKanjiFilterRules) {
-        store.put(rule);
-      }
-    },
-  });
-  const filterRules = await db.getAll(DATABASE.onlyTable);
+  const db = await getKanjiFilterDB();
+  const filterRules = await db.getAll(DB.onlyTable);
   const filterMap = new Map<string, string[]>(
     filterRules.map((filterRule) => [filterRule.kanji, filterRule.reading]),
   );
